handle overflows properly for stack increment and arithmetic add operations

+++ src/ops/op_handlers/arithmetic.rs
@@ -12,15 +12,15 @@ pub fn add_handler(
     let mut wrapper = UxnWrapper::new(u, keep, ret);
 
     if short == true {
-        let a = wrapper.pop_short()?;
-        let b = wrapper.pop_short()?;
+        let a = wrapper.pop_short()? as i32;
+        let b = wrapper.pop_short()? as i32;
 
-        wrapper.push_short(a+b)?;
+        wrapper.push_short((a+b) as u16)?;
     } else {
-        let a = wrapper.pop()?;
-        let b = wrapper.pop()?;
+        let a = wrapper.pop()? as i16;
+        let b = wrapper.pop()? as i16;
 
-        wrapper.push(a+b)?;
+        wrapper.push((a+b) as u8)?;
     }
 
     return Ok(());
